+++ backend/poskds/src/main/java/com/anyawalker/poskds/repos/OrderItemRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderItemRepo extends JpaRepository<@NonNull OrderItemEntity,@NonNull Long> {
+public interface OrderItemRepo extends JpaRepository<@NonNull OrderItemEntity,@NonNull Integer> {
 }
